Move facets rendering to a separate class
